feat: home 컴포넌트 section별로 분리
